After jesse's code review, lots to clean up hehe Remove hasher worker completely and incorporate its functions into StateHistoryReader and TEVMIndexer Refactor most helper functions related to evm hashing to utils/evm.ts Remove gaps checking code for now Remove fork handling code for now Add better logging & more debug logging
